+++ packages/views/locales/en/chat.json
@@ -27,9 +27,6 @@
     "copy_failed_toast": "Copy failed"
   },
   "session_history": {
-    "back_tooltip": "Back",
-    "header": "Chat History",
-    "empty": "No chat sessions yet",
     "untitled": "Untitled",
     "time": {
       "just_now": "just now",
@@ -38,7 +35,6 @@
       "days": "{{count}}d ago"
     },
     "row_delete_aria": "Delete chat session",
-    "row_delete_tooltip": "Delete",
     "delete_dialog": {
       "title": "Delete chat session",
       "description_with_title": "\"{{title}}\" and its messages will be permanently removed. This action cannot be undone.",
@@ -62,8 +58,9 @@
     "my_agents": "My agents",
     "others": "Others",
     "no_agents": "No agents",
-    "history_show_tooltip": "Chat history",
-    "history_back_tooltip": "Back to chat"
+    "active_group": "Active",
+    "archived_group_one": "{{count}} archived chat",
+    "archived_group_other": "{{count}} archived chats"
   },
   "empty_state": {
     "first_time_title": "Chat with your agents",
